Extract sync regenerator parts
