+++ packages/babel-preset-env/test/fixtures/corejs2-babel-7/usage-regenerator-used-async/output.mjs
@@ -1,27 +1,23 @@
 function _typeof(o) { "@babel/helpers - typeof"; return _typeof = "function" == typeof Symbol && "symbol" == typeof Symbol.iterator ? function (o) { return typeof o; } : function (o) { return o && "function" == typeof Symbol && o.constructor === Symbol && o !== Symbol.prototype ? "symbol" : typeof o; }, _typeof(o); }
-function _regeneratorRuntime() { "use strict"; /*! regenerator-runtime -- Copyright (c) 2014-present, Facebook, Inc. -- license (MIT): https://github.com/babel/babel/blob/main/packages/babel-helpers/LICENSE */ _regeneratorRuntime = function _regeneratorRuntime() { return r; }; var t, r = { keys: _regeneratorKeys, awrap: _awaitAsyncGenerator, async: _regeneratorAsync, AsyncIterator: _regeneratorAsyncIterator }, e = Object.prototype, n = e.hasOwnProperty, o = "function" == typeof Symbol ? Symbol : {}, i = o.iterator || "@@iterator", a = o.toStringTag || "@@toStringTag"; r.wrap = function (r, e, n, o) { var i = e && e.prototype instanceof Generator ? e : Generator, a = Object.create(i.prototype), h = new Context(o || []); return _regeneratorDefine(a, "_invoke", function (r, e, n) { var o = 1; return function (i, a) { if (3 === o) throw Error("Generator is already running"); if (4 === o) { if ("throw" === i) throw a; return { value: t, done: !0 }; } for (n.method = i, n.arg = a;;) { var h = n.delegate; if (h) { var s = p(h, n); if (s) { if (s === u) continue; return s; } } if ("next" === n.method) n.sent = n._sent = n.arg;else if ("throw" === n.method) { if (1 === o) throw o = 4, n.arg; n.dispatchException(n.arg); } else "return" === n.method && n.abrupt("return", n.arg); o = 3; var c = _tryCatch(r, e, n); if (!c.e) { if (o = n.done ? 4 : 2, c.v === u) continue; return { value: c.v, done: n.done }; } o = 4, n.method = "throw", n.arg = c.v; } }; }(r, n, h), !0), a; }; var u = {}; function Generator() {} function GeneratorFunction() {} function GeneratorFunctionPrototype() {} var h = {}; _regeneratorDefine(h, i, function () { return this; }); var s = Object.getPrototypeOf, c = s && s(s(d([]))); c && c !== e && n.call(c, i) && (h = c); var f = GeneratorFunctionPrototype.prototype = Generator.prototype = Object.create(h); function p(r, e) { var n = e.method, o = r.i[n]; if (o === t) return e.delegate = null, "throw" === n && r.i["return"] && (e.method = "return", e.arg = t, p(r, e), "throw" === e.method) || "return" !== n && (e.method = "throw", e.arg = new TypeError("The iterator does not provide a '" + n + "' method")), u; var i = _tryCatch(o, r.i, e.arg); if (i.e) return e.method = "throw", e.arg = i.v, e.delegate = null, u; var a = i.v; return a ? a.done ? (e[r.r] = a.value, e.next = r.n, "return" !== e.method && (e.method = "next", e.arg = t), e.delegate = null, u) : a : (e.method = "throw", e.arg = new TypeError("iterator result is not an object"), e.delegate = null, u); } function y(t) { this.tryEntries.push(t); } function l(r) { var e = r[4] || {}; e.type = "normal", e.arg = t, r[4] = e; } function Context(t) { this.tryEntries = [[-1]], t.forEach(y, this), this.reset(!0); } function d(r) { if (null != r) { var e = r[i]; if (e) return e.call(r); if ("function" == typeof r.next) return r; if (!isNaN(r.length)) { var o = -1, a = function e() { for (; ++o < r.length;) if (n.call(r, o)) return e.value = r[o], e.done = !1, e; return e.value = t, e.done = !0, e; }; return a.next = a; } } throw new TypeError(_typeof(r) + " is not iterable"); } return GeneratorFunction.prototype = GeneratorFunctionPrototype, _regeneratorDefine(f, "constructor", GeneratorFunctionPrototype), _regeneratorDefine(GeneratorFunctionPrototype, "constructor", GeneratorFunction), GeneratorFunction.displayName = _regeneratorDefine(GeneratorFunctionPrototype, a, "GeneratorFunction"), r.isGeneratorFunction = function (t) { var r = "function" == typeof t && t.constructor; return !!r && (r === GeneratorFunction || "GeneratorFunction" === (r.displayName || r.name)); }, r.mark = function (t) { return Object.setPrototypeOf ? Object.setPrototypeOf(t, GeneratorFunctionPrototype) : (t.__proto__ = GeneratorFunctionPrototype, _regeneratorDefine(t, a, "GeneratorFunction")), t.prototype = Object.create(f), t; }, _regeneratorDefineIM(f), _regeneratorDefine(f, a, "Generator"), _regeneratorDefine(f, i, function () { return this; }), _regeneratorDefine(f, "toString", function () { return "[object Generator]"; }), r.values = d, Context.prototype = { constructor: Context, reset: function reset(r) { if (this.prev = this.next = 0, this.sent = this._sent = t, this.done = !1, this.delegate = null, this.method = "next", this.arg = t, this.tryEntries.forEach(l), !r) for (var e in this) "t" === e.charAt(0) && n.call(this, e) && !isNaN(+e.slice(1)) && (this[e] = t); }, stop: function stop() { this.done = !0; var t = this.tryEntries[0][4]; if ("throw" === t.type) throw t.arg; return this.rval; }, dispatchException: function dispatchException(r) { if (this.done) throw r; var e = this; function n(t) { a.type = "throw", a.arg = r, e.next = t; } for (var o = e.tryEntries.length - 1; o >= 0; --o) { var i = this.tryEntries[o], a = i[4], u = this.prev, h = i[1], s = i[2]; if (-1 === i[0]) return n("end"), !1; if (!h && !s) throw Error("try statement without catch or finally"); if (null != i[0] && i[0] <= u) { if (u < h) return this.method = "next", this.arg = t, n(h), !0; if (u < s) return n(s), !1; } } }, abrupt: function abrupt(t, r) { for (var e = this.tryEntries.length - 1; e >= 0; --e) { var n = this.tryEntries[e]; if (n[0] > -1 && n[0] <= this.prev && this.prev < n[2]) { var o = n; break; } } o && ("break" === t || "continue" === t) && o[0] <= r && r <= o[2] && (o = null); var i = o ? o[4] : {}; return i.type = t, i.arg = r, o ? (this.method = "next", this.next = o[2], u) : this.complete(i); }, complete: function complete(t, r) { if ("throw" === t.type) throw t.arg; return "break" === t.type || "continue" === t.type ? this.next = t.arg : "return" === t.type ? (this.rval = this.arg = t.arg, this.method = "return", this.next = "end") : "normal" === t.type && r && (this.next = r), u; }, finish: function finish(t) { for (var r = this.tryEntries.length - 1; r >= 0; --r) { var e = this.tryEntries[r]; if (e[2] === t) return this.complete(e[4], e[3]), l(e), u; } }, "catch": function _catch(t) { for (var r = this.tryEntries.length - 1; r >= 0; --r) { var e = this.tryEntries[r]; if (e[0] === t) { var n = e[4]; if ("throw" === n.type) { var o = n.arg; l(e); } return o; } } throw Error("illegal catch attempt"); }, delegateYield: function delegateYield(r, e, n) { return this.delegate = { i: d(r), r: e, n: n }, "next" === this.method && (this.arg = t), u; } }, r; }
-function _regeneratorAsync(r, e, n, t, o) { null == o && (o = Promise); var a = _regeneratorRuntime(), u = new _regeneratorAsyncIterator(a.wrap(r, e, n, t), o); return a.isGeneratorFunction(e) ? u : u.next().then(function (r) { return r.done ? r.value : u.next(); }); }
-function _regeneratorAsyncIterator(e, t) { function invoke(n, o, r, i) { var a = _tryCatch(e[n], e, o); if (!a.e) { var f = a.v, c = f.value; return c && c instanceof _OverloadYield ? t.resolve(c.v).then(function (e) { invoke("next", e, r, i); }, function (e) { invoke("throw", e, r, i); }) : t.resolve(c).then(function (e) { f.value = e, r(f); }, function (e) { return invoke("throw", e, r, i); }); } i(a.v); } var n; this.next || (_regeneratorDefineIM(_regeneratorAsyncIterator.prototype), _regeneratorDefine(_regeneratorAsyncIterator.prototype, "function" == typeof Symbol && Symbol.asyncIterator || "@asyncIterator", function () { return this; })), _regeneratorDefine(this, "_invoke", function (e, o) { function callInvokeWithMethodAndArg() { return new t(function (t, n) { invoke(e, o, t, n); }); } return n = n ? n.then(callInvokeWithMethodAndArg, callInvokeWithMethodAndArg) : callInvokeWithMethodAndArg(); }, !0); }
-function _regeneratorDefineIM(e) { ["next", "throw", "return"].forEach(function (n) { _regeneratorDefine(e, n, function (e) { return this._invoke(n, e); }); }); }
+function _regeneratorRuntime() { "use strict"; var r = _regenerator(); return (_regeneratorRuntime = function _regeneratorRuntime() { return { wrap: r.w, mark: r.m, values: _regeneratorValues, isGeneratorFunction: function isGeneratorFunction(e) { var t = "function" == typeof e && e.constructor; return !!t && (t === r.G || "GeneratorFunction" === (t.displayName || t.name)); } }; })(); }
+function _regenerator() { "use strict"; /*! regenerator-runtime -- Copyright (c) 2014-present, Facebook, Inc. -- license (MIT): https://github.com/babel/babel/blob/main/packages/babel-helpers/LICENSE */ var t, r = Object.prototype, e = r.hasOwnProperty, n = "function" == typeof Symbol ? Symbol : {}, o = n.iterator || "@@iterator", i = n.toStringTag || "@@toStringTag"; function a(r, e, n, o) { var i = e && e.prototype instanceof Generator ? e : Generator, a = Object.create(i.prototype), u = new Context(o || []); return _regeneratorDefine(a, "_invoke", function (r, e, n) { var o = 1; return function (i, a) { if (3 === o) throw Error("Generator is already running"); if (4 === o) { if ("throw" === i) throw a; return { value: t, done: !0 }; } for (n.method = i, n.arg = a;;) { var u = n.delegate; if (u) { var s = d(u, n); if (s) { if (s === h) continue; return s; } } if ("next" === n.method) n.sent = n._sent = n.arg;else if ("throw" === n.method) { if (1 === o) throw o = 4, n.arg; n.dispatchException(n.arg); } else "return" === n.method && n.abrupt("return", n.arg); o = 3; var c = _tryCatch(r, e, n); if (!c.e) { if (o = n.done ? 4 : 2, c.v === h) continue; return { value: c.v, done: n.done }; } o = 4, n.method = "throw", n.arg = c.v; } }; }(r, n, u), !0), a; } var h = {}; function Generator() {} function GeneratorFunction() {} function GeneratorFunctionPrototype() {} var u = {}; _regeneratorDefine(u, o, function () { return this; }); var s = Object.getPrototypeOf, c = s && s(s(_regeneratorValues([]))); c && c !== r && e.call(c, o) && (u = c); var f = GeneratorFunctionPrototype.prototype = Generator.prototype = Object.create(u); function p(t) { return Object.setPrototypeOf ? Object.setPrototypeOf(t, GeneratorFunctionPrototype) : (t.__proto__ = GeneratorFunctionPrototype, _regeneratorDefine(t, i, "GeneratorFunction")), t.prototype = Object.create(f), t; } function d(r, e) { var n = e.method, o = r.i[n]; if (o === t) return e.delegate = null, "throw" === n && r.i["return"] && (e.method = "return", e.arg = t, d(r, e), "throw" === e.method) || "return" !== n && (e.method = "throw", e.arg = new TypeError("The iterator does not provide a '" + n + "' method")), h; var i = _tryCatch(o, r.i, e.arg); if (i.e) return e.method = "throw", e.arg = i.v, e.delegate = null, h; var a = i.v; return a ? a.done ? (e[r.r] = a.value, e.next = r.n, "return" !== e.method && (e.method = "next", e.arg = t), e.delegate = null, h) : a : (e.method = "throw", e.arg = new TypeError("iterator result is not an object"), e.delegate = null, h); } function l(t) { this.tryEntries.push(t); } function y(r) { var e = r[4] || {}; e.type = "normal", e.arg = t, r[4] = e; } function Context(t) { this.tryEntries = [[-1]], t.forEach(l, this), this.reset(!0); } return GeneratorFunction.prototype = GeneratorFunctionPrototype, _regeneratorDefine(f, "constructor", GeneratorFunctionPrototype), _regeneratorDefine(GeneratorFunctionPrototype, "constructor", GeneratorFunction), GeneratorFunction.displayName = _regeneratorDefine(GeneratorFunctionPrototype, i, "GeneratorFunction"), _regeneratorDefineIM(f), _regeneratorDefine(f, i, "Generator"), _regeneratorDefine(f, o, function () { return this; }), _regeneratorDefine(f, "toString", function () { return "[object Generator]"; }), Context.prototype = { constructor: Context, reset: function reset(r) { if (this.prev = this.next = 0, this.sent = this._sent = t, this.done = !1, this.delegate = null, this.method = "next", this.arg = t, this.tryEntries.forEach(y), !r) for (var n in this) "t" === n.charAt(0) && e.call(this, n) && !isNaN(+n.slice(1)) && (this[n] = t); }, stop: function stop() { this.done = !0; var t = this.tryEntries[0][4]; if ("throw" === t.type) throw t.arg; return this.rval; }, dispatchException: function dispatchException(r) { if (this.done) throw r; var e = this; function n(t) { a.type = "throw", a.arg = r, e.next = t; } for (var o = e.tryEntries.length - 1; o >= 0; --o) { var i = this.tryEntries[o], a = i[4], h = this.prev, u = i[1], s = i[2]; if (-1 === i[0]) return n("end"), !1; if (!u && !s) throw Error("try statement without catch or finally"); if (null != i[0] && i[0] <= h) { if (h < u) return this.method = "next", this.arg = t, n(u), !0; if (h < s) return n(s), !1; } } }, abrupt: function abrupt(t, r) { for (var e = this.tryEntries.length - 1; e >= 0; --e) { var n = this.tryEntries[e]; if (n[0] > -1 && n[0] <= this.prev && this.prev < n[2]) { var o = n; break; } } o && ("break" === t || "continue" === t) && o[0] <= r && r <= o[2] && (o = null); var i = o ? o[4] : {}; return i.type = t, i.arg = r, o ? (this.method = "next", this.next = o[2], h) : this.complete(i); }, complete: function complete(t, r) { if ("throw" === t.type) throw t.arg; return "break" === t.type || "continue" === t.type ? this.next = t.arg : "return" === t.type ? (this.rval = this.arg = t.arg, this.method = "return", this.next = "end") : "normal" === t.type && r && (this.next = r), h; }, finish: function finish(t) { for (var r = this.tryEntries.length - 1; r >= 0; --r) { var e = this.tryEntries[r]; if (e[2] === t) return this.complete(e[4], e[3]), y(e), h; } }, "catch": function _catch(t) { for (var r = this.tryEntries.length - 1; r >= 0; --r) { var e = this.tryEntries[r]; if (e[0] === t) { var n = e[4]; if ("throw" === n.type) { var o = n.arg; y(e); } return o; } } throw Error("illegal catch attempt"); }, delegateYield: function delegateYield(r, e, n) { return this.delegate = { i: _regeneratorValues(r), r: e, n: n }, "next" === this.method && (this.arg = t), h; } }, (_regenerator = function _regenerator() { return { w: a, m: p, G: GeneratorFunction }; })(); }
 function _tryCatch(t, r, e) { try { return { e: 0, v: t.call(r, e) }; } catch (t) { return { e: 1, v: t }; } }
+function _regeneratorValues(e) { if (null != e) { var t = e["function" == typeof Symbol && Symbol.iterator || "@iterator"]; if (t) return t.call(e); if ("function" == typeof e.next) return e; if (!isNaN(e.length)) { var n = -1, r = function next() { for (; ++n < e.length;) if ({}.hasOwnProperty.call(e, n)) return next.value = e[n], next.done = !1, next; return next.value = void 0, next.done = !0, next; }; return r.next = r; } } throw new TypeError(_typeof(e) + " is not iterable"); }
+function _regeneratorDefineIM(e) { ["next", "throw", "return"].forEach(function (n) { _regeneratorDefine(e, n, function (e) { return this._invoke(n, e); }); }); }
 function _regeneratorDefine(e, r, n, t) { _regeneratorDefine = function _regeneratorDefine(e, r, n, t) { return Object.defineProperty(e, r, { value: n, enumerable: !t, configurable: !t, writable: !t }); }; try { _regeneratorDefine({}, ""); } catch (e) { _regeneratorDefine = function _regeneratorDefine(e, r, n) { return e[r] = n; }; } return _regeneratorDefine(e, r, n, t); }
-function _regeneratorKeys(e) { var n = Object(e), t = []; for (var r in n) t.unshift(r); return function next() { for (; t.length;) if ((r = t.pop()) in n) return next.value = r, next.done = !1, next; return next.done = !0, next; }; }
-function _awaitAsyncGenerator(e) { return new _OverloadYield(e, 0); }
-function _OverloadYield(e, d) { this.v = e, this.k = d; }
 import "core-js/modules/es6.object.to-string.js";
 import "core-js/modules/es6.promise.js";
 import "core-js/modules/es6.object.define-property.js";
 import "core-js/modules/es6.array.for-each.js";
 import "core-js/modules/es6.symbol.js";
-import "core-js/modules/es7.symbol.async-iterator.js";
-import "core-js/modules/es6.object.create.js";
-import "core-js/modules/es6.object.get-prototype-of.js";
-import "core-js/modules/es6.function.name.js";
-import "core-js/modules/es6.object.set-prototype-of.js";
-import "core-js/modules/es6.array.slice.js";
 import "core-js/modules/es6.string.iterator.js";
 import "core-js/modules/es6.array.iterator.js";
 import "core-js/modules/web.dom.iterable.js";
+import "core-js/modules/es6.object.create.js";
+import "core-js/modules/es6.object.get-prototype-of.js";
+import "core-js/modules/es6.object.set-prototype-of.js";
+import "core-js/modules/es6.array.slice.js";
+import "core-js/modules/es6.function.name.js";
 function asyncGeneratorStep(n, t, e, r, o, a, c) { try { var i = n[a](c), u = i.value; } catch (n) { return void e(n); } i.done ? t(u) : Promise.resolve(u).then(r, o); }
 function _asyncToGenerator(n) { return function () { var t = this, e = arguments; return new Promise(function (r, o) { var a = n.apply(t, e); function _next(n) { asyncGeneratorStep(a, r, o, _next, _throw, "next", n); } function _throw(n) { asyncGeneratorStep(a, r, o, _next, _throw, "throw", n); } _next(void 0); }); }; }
 function a() {
